% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_genelist.R
\name{annotate_biomart}
\alias{annotate_biomart}
\title{Annotate Biomart}
\usage{
annotate_biomart(res, gm)
}
\arguments{
\item{res}{DESeqResults object, must have ensembl gene ID in 'GeneID' column}

\item{gm}{biomart genemap, contains GeneID and MGI_symbol columns}
}
\value{
results object now with annotations
}
\description{
add human readable gene IDs, descriptions, and gene biotypes to the data
}
