% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rank_export_DEGs.R
\name{export_genelist}
\alias{export_genelist}
\title{Export Gene List}
\usage{
export_genelist(tab, filename, direction = NULL, saveas = ".csv")
}
\arguments{
\item{tab}{data frame, contains ranked DEGs, the output of signif_deg}

\item{filename}{string, name to save the exported file as}

\item{direction}{string ("up", "down", or NULL), which genes to export}

\item{saveas}{string, extention for filename, defaults to ".csv"}
}
\value{
CSV file exported to CWD
}
\description{
Export Gene List
}
