% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.r
\name{get_DEresults}
\alias{get_DEresults}
\title{Get DE results}
\usage{
get_DEresults(
  contrast_str = "Condition",
  numerator,
  denominator,
  tidy_result = TRUE,
  deseq_obj
)
}
\arguments{
\item{contrast_str}{string, defaults to "Condition", the comparison of interest for pairwise comparison}

\item{numerator}{string, name of the conditon of interest for comparison}

\item{denominator}{string, name of baseline condition for comparison}

\item{tidy_result}{boolean, defaults to TRUE, dictates if the results are returned as a dataframe or obj}

\item{deseq_obj}{data}
}
\value{
DESeq Results object, or a dataframe
}
\description{
Get the results of a deseq object
}
