% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{pca_analysis}
\alias{pca_analysis}
\title{PCA Analysis}
\usage{
pca_analysis(
  dseq_transform,
  labels,
  colors,
  shapes,
  plot_title = "Principal Component Analysis",
  label_samples = TRUE
)
}
\arguments{
\item{dseq_transform}{DESeqTransform object, either VST or RLog transformed}

\item{labels}{list, conditions being plotted}

\item{colors}{list, colors for each group}

\item{shapes}{list, shapes for each group}

\item{plot_title}{string, defaults to "Principal Component Analysis"}

\item{label_samples}{boolean, if true, labels for each sample will be added, defaults to TRUE}
}
\value{
PCA plot
}
\description{
Illustrate PCA plot to show sample relationships
}
