% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genecounts.R
\name{plot_genecounts}
\alias{plot_genecounts}
\title{Plot Gene Counts}
\usage{
plot_genecounts(
  geneid,
  dseq_obj,
  plot_colors,
  plot_shapes,
  plot_labels,
  plot_title = "Gene Counts",
  plot_subtitle = "",
  groupby = "condition",
  label_samples = TRUE
)
}
\arguments{
\item{geneid}{string, Ensembl Gene ID}

\item{dseq_obj}{DESeq object}

\item{plot_title}{string, plot title, defaults to "Gene Counts"}

\item{plot_subtitle}{string, subtitle for plot, defaults to empty string}

\item{groupby}{string, interest group, defaults to "Condition"}

\item{label_samples}{boolean, indicates whether samples should be labeled on plot, defaults to TRUE}

\item{colors}{list, colors to use for each condition}

\item{shapes}{list, shapes to use for each condition}

\item{labels}{list, labels to use for each condition}
}
\value{
image
}
\description{
Plot the gene counts of a certain gene of interest
}
