% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_genelist.R
\name{rename_columns}
\alias{rename_columns}
\title{Rename Columns}
\usage{
rename_columns(colstr)
}
\arguments{
\item{colstr}{string, column to be renamed}
}
\value{
string, edited column name
}
\description{
rename columns, this allows for consistent naming and easier access to data
}
