% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rank_export_DEGs.R
\name{signif_deg}
\alias{signif_deg}
\title{Filter and Rank Significant Differentially Expressed Genes}
\usage{
signif_deg(result, padj_cutoff = 0.05, l2fc_cutoff = 0.8)
}
\arguments{
\item{result}{dataframe contains \itemize{
\item gene annotations
\item log2 FC and adjusted p value}}

\item{padj_cutoff}{double, cutoff level for adjusted P vals to keep, defaults to 0.05}

\item{l2fc_cutoff}{double, cutoff level for log2 fold change, defaults to 0.8 (abs value)}
}
\value{
ranked, a sorted and filtered data frame containing \itemize{
\item annotations
\item change direction
\item fold change
\item adjusted p value
\item change metric (log2 fold change * - log10 padj)}
}
\description{
This function \itemize{
\item keeps only differentially expressed genes that are significant
\item adds an indicator of if genes in the experimental group are upregulated or downregulated compared to baseline
\item and calculates a change metric which will combine direction, Log2 Fold Change, and adjusted Pvalue}
}
