% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{umap_analysis}
\alias{umap_analysis}
\title{UMAP Analysis}
\usage{
umap_analysis(
  dseq_transform,
  neighbors,
  m_dist,
  cond_list,
  plot_colors,
  plot_shapes,
  plot_title = "UMAP",
  label_samples = TRUE,
  batch = FALSE
)
}
\arguments{
\item{dseq_transform}{DESeqTransform object, either VST or RLog transformed}

\item{neighbors}{integer, minimum of 2}

\item{m_dist}{integer}

\item{cond_list}{list, conditions for each group}

\item{plot_colors}{list (strings), colors for each group}

\item{plot_shapes}{list (integers), shapes for each group}

\item{plot_title}{string, plot title, defaults to "UMAP"}

\item{label_samples}{boolean, if sample labels should be added to plot, defaults to true}

\item{batch}{boolean, defaults to false, set to true & add set seed for reproducible results}
}
\value{
UMAP plot
}
\description{
plot umap for all samples, reducing the number of variables and showing relationships between samples
}
